data_encrypter/requirements.txt
pycrypto
termcolor
argparse
glob2
customtkinter
hashlib
pyotp